+++ examples.config.js
@@ -26,6 +26,11 @@ module.exports = {
             compressor: {
                 warnings: false
             }
+        }),
+        new webpack.DefinePlugin({
+            "process.env": {
+                "NODE_ENV": JSON.stringify("production")
+            }
         })
     ]
 };