+++ server.js
@@ -9,10 +9,16 @@ var express = require("express"),
 config.module.loaders[0].query = {presets: ["react-hmre"]};
 config.entry.unshift("webpack-hot-middleware/client");
 
-config.plugins.push(
+config.plugins = [
+    new webpack.optimize.OccurenceOrderPlugin(),
+    new webpack.optimize.UglifyJsPlugin({
+        compressor: {
+            warnings: false
+        }
+    }),
     new webpack.HotModuleReplacementPlugin(),
     new webpack.NoErrorsPlugin()
-);
+];
 
 var app = express();