+++ src/redux/store.js
@@ -1,0 +1,6 @@
+"use strict";
+
+import { createStore } from "redux";
+import rootReducer from "./reducers";
+
+export default createStore(rootReducer);